requirements.txt
asgiref==3.2.7
bson==0.5.8
config==0.5.0.post0
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.13
django-cors-headers==3.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djongo==1.3.2
Pillow==7.1.2
psycopg2==2.8.5
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
whitenoise==5.1.0
